use QDoubleSpinBox instead of the integer version for free space before and after and for capacity so that fractions are possible for these values
svn path=/trunk/extragear/sysadmin/partitionmanager/; revision=1118931

+++ src/gui/sizedialogbase.h
@@ -74,9 +74,9 @@ class SizeDialogBase : public KDialog
 		void onFirstSectorChanged(qint64 newFirst);
 		void onLastSectorChanged(qint64 newLast);
 
-		void onCapacityChanged(int newCapacity);
-		void onFreeSpaceBeforeChanged(int newBefore);
-		void onFreeSpaceAfterChanged(int newAfter);
+		void onCapacityChanged(double newCapacity);
+		void onFreeSpaceBeforeChanged(double newBefore);
+		void onFreeSpaceAfterChanged(double newAfter);
 
 		void onSpinFirstSectorChanged(double newFirst);
 		void onSpinLastSectorChanged(double newLast);
